# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Titre de votre API",
        "description": "Description de votre API",
        "contact": {
            "email": "support@votreapi.com"
        },
        "license": {
            "name": "Licence de votre API",
            "url": "URL de la licence"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "AllUsersWithOptionalFilters"
                ],
                "summary": "Get all users with optional filters",
                "description": "Récupération des utilisateurs avec possibilité de filtrer par role et/ou par active",
                "operationId": "GetAllUsersWithOptionalFilters",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filterer les utilisateurs par role (admin ou boutiquier)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "boutiquier"
                            ]
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filterer les utilisateurs par status (oui ou non)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "oui",
                                "non"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un nouvel utilisateur",
                "description": "Crée un nouvel utilisateur avec les données fournies dans la requête.",
                "operationId": "ba548b97148bb5c0d8cab3bc3f128bfd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Afficher un utilisateur spécifique",
                "description": "Récupère les informations d'un utilisateur spécifique par son ID.",
                "operationId": "daaf7a6169ee9348c3fddb07778598ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "En-tête d'autorisation JWT utilisant le schéma Bearer."
            }
        }
    },
    "security": [
        {
            "BearerToken": []
        }
    ]
}